feat: split NavUserMenu into server shell + NavUserMenuClient (#11)
